// rnsis-nexus/packages/shared/src/enums.ts
export const LateFeeMode = {
  FIXED: 'FIXED',
  PERCENT: 'PERCENT',
} as const;
export type LateFeeMode = (typeof LateFeeMode)[keyof typeof LateFeeMode];
export const LateFeeModeValues = Object.values(LateFeeMode) as LateFeeMode[];

export const LateFeeFrequency = {
  ONCE: 'ONCE',
  PER_DAY: 'PER_DAY',
  PER_WEEK: 'PER_WEEK',
  PER_MONTH: 'PER_MONTH',
} as const;
export type LateFeeFrequency = (typeof LateFeeFrequency)[keyof typeof LateFeeFrequency];
// ...
export const DiscountMode = {
  PERCENT: 'PERCENT',
  FIXED: 'FIXED',
} as const;
export type DiscountMode = (typeof DiscountMode)[keyof typeof DiscountMode];
// ...
export const InvoiceStatus = {
  PENDING: 'PENDING',
  PARTIALLY_PAID: 'PARTIALLY_PAID',
  PAID: 'PAID',
  OVERDUE: 'OVERDUE',
  CANCELLED: 'CANCELLED',
} as const;
export type InvoiceStatus = (typeof InvoiceStatus)[keyof typeof InvoiceStatus];